+++ Helper/ProductFilters.php
@@ -129,24 +129,10 @@ class ProductFilters
                     /** @var string[] $familyFilter */
                     $familyFilter = ['operator' => 'IN', 'value' => [$productFamily]];
                     $advancedFilters['search']['family'][] = $familyFilter;
-                    $productFilterAdded = true;
                 }
             }
 
-            if (!empty($advancedFilters['scope'])) {
-                if (!in_array($advancedFilters['scope'], $mappedChannels)) {
-                    /** @var string[] $error */
-                    $error = [
-                        'error' => __('Advanced filters contains an unauthorized scope, please add check your filters and website mapping.'),
-                    ];
-
-                    return $error;
-                }
-
-                return [$advancedFilters];
-            }
-
-            $search = $advancedFilters['search'];
+            return [$advancedFilters];
         }
 
         if ($mode == Mode::STANDARD) {
@@ -159,10 +145,9 @@ class ProductFilters
         }
 
         // If import product gave a family, add this family to the search
-        if ($productFamily && !$productFilterAdded) {
+        if ($productFamily) {
             $familyFilter = ['operator' => 'IN', 'value' => [$productFamily]];
             $search['family'][] = $familyFilter;
-            $productFilterAdded = true;
         }
 
         /** @var string $channel */
@@ -172,12 +157,6 @@ class ProductFilters
                 'search' => $search,
                 'scope'  => $channel,
             ];
-
-            if ($mode == Mode::ADVANCED) {
-                $filters[] = $filter;
-
-                continue;
-            }
 
             if ($this->configHelper->getCompletenessTypeFilter() !== Completeness::NO_CONDITION) {
                 /** @var string[] $completeness */
@@ -190,6 +169,10 @@ class ProductFilters
                 }
             }
 
+            if ($this->configHelper->getAttributeFilterByCodeMode() == true) {
+                $filter['attributes'] = $this->configHelper->getAttributeFilterByCode();
+            }
+            
             /** @var string[] $locales */
             $locales = $this->storeHelper->getChannelStoreLangs($channel);
             if (!empty($locales)) {